+++ semgrep/semgrep/output.py
@@ -77,12 +77,14 @@ def build_normal_output(
         RESET_COLOR = colorama.Style.RESET_ALL if color_output else ""
         GREEN_COLOR = colorama.Fore.GREEN if color_output else ""
         YELLOW_COLOR = colorama.Fore.YELLOW if color_output else ""
+        RED_COLOR = colorama.Fore.RED if color_output else ""
         BLUE_COLOR = colorama.Fore.BLUE if color_output else ""
 
         current_file = finding.get("path", "<no path>")
         check_id = finding.get("check_id")
         extra = finding.get("extra", {})
         message = extra.get("message")
+        severity = extra.get("severity")
         fix = extra.get("fix")
         if last_file is None or last_file != current_file:
             if last_file is not None:
@@ -95,7 +97,15 @@ def build_normal_output(
             and check_id != "-"
             and (last_message is None or last_message != message)
         ):
-            yield f"{YELLOW_COLOR}rule:{check_id}: {finding.get('extra', {}).get('message')}{RESET_COLOR}"
+            severity_prepend = ""
+            if severity:
+                if severity == "ERROR":
+                    severity_prepend = f"{RED_COLOR}{severity} "
+                elif severity == "WARNING":
+                    severity_prepend = f"{YELLOW_COLOR}{severity} "
+                else:
+                    severity_prepend = f"{severity} "
+            yield f"{severity_prepend}{YELLOW_COLOR}rule:{check_id}: {finding.get('extra', {}).get('message')}{RESET_COLOR}"
 
         last_file = current_file
         last_message = message